+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ latest
 
 * Recursive wildcard support for ignored imports.
 * Drop support for Python 3.7.
+* Use grimp.ImportGraph instead of importlinter.domain.ports.graph.ImportGraph.
 
 1.9.0 (2023-05-13)
 ------------------